requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests

chromadb
pypdf
google-generativeai